// graphql/resolvers.d.ts
export type Maybe<T> = T | null;
// ...
export type RequireFields<T, K extends keyof T> = Omit<T, K> & { [P in K]-?: NonNullable<T[P]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
    ID: { input: string; output: string };
    String: { input: string; output: string };
    Boolean: { input: boolean; output: boolean };
    Int: { input: number; output: number };
    Float: { input: number; output: number };
};

export type Query = {
    __typename?: 'Query';
    user?: Maybe<User>;
};

export type QueryUserArgs = {
    name: Scalars['String']['input'];
};

export enum Role {
    Admin = 'ADMIN',
    User = 'USER',
}

export type User = {
    __typename?: 'User';
    email?: Maybe<Scalars['String']['output']>;
    name?: Maybe<Scalars['String']['output']>;
    role?: Maybe<Role>;
};

export type UserWithPassword = {
    __typename?: 'UserWithPassword';
    email?: Maybe<Scalars['String']['output']>;
    name?: Maybe<Scalars['String']['output']>;
    password?: Maybe<Scalars['String']['output']>;
    role?: Maybe<Role>;
};

export type WithIndex<TObject> = TObject & Record<string, any>;
export type ResolversObject<TObject> = WithIndex<TObject>;

export type ResolverTypeWrapper<T> = Promise<T> | T;

export type ResolverWithResolve<TResult, TParent, TContext, TArgs> = {
    resolve: ResolverFn<TResult, TParent, TContext, TArgs>;
};
export type Resolver<TResult, TParent = {}, TContext = {}, TArgs = {}> =
    | ResolverFn<TResult, TParent, TContext, TArgs>
    | ResolverWithResolve<TResult, TParent, TContext, TArgs>;

export type ResolverFn<TResult, TParent, TContext, TArgs> = (
    parent: TParent,
    args: TArgs,
    context: TContext,
    info: GraphQLResolveInfo,
) => Promise<TResult> | TResult;
// ...
export type IsTypeOfResolverFn<T = {}, TContext = {}> = (
    obj: T,
    context: TContext,
    info: GraphQLResolveInfo,
) => boolean | Promise<boolean>;

export type NextResolverFn<T> = () => Promise<T>;

export type DirectiveResolverFn<TResult = {}, TParent = {}, TContext = {}, TArgs = {}> = (
    next: NextResolverFn<TResult>,
    parent: TParent,
    args: TArgs,
    context: TContext,
    info: GraphQLResolveInfo,
) => TResult | Promise<TResult>;

/** Mapping between all available schema types and the resolvers types */
export type ResolversTypes = ResolversObject<{
    Boolean: ResolverTypeWrapper<Scalars['Boolean']['output']>;
    Query: ResolverTypeWrapper<{}>;
    Role: Role;
    String: ResolverTypeWrapper<Scalars['String']['output']>;
    User: ResolverTypeWrapper<User>;
    UserWithPassword: ResolverTypeWrapper<UserWithPassword>;
}>;

/** Mapping between all available schema types and the resolvers parents */
export type ResolversParentTypes = ResolversObject<{
    Boolean: Scalars['Boolean']['output'];
    Query: {};
    String: Scalars['String']['output'];
    User: User;
    UserWithPassword: UserWithPassword;
}>;

export type QueryResolvers<
    ContextType = any,
    ParentType extends ResolversParentTypes['Query'] = ResolversParentTypes['Query'],
> = ResolversObject<{
    user?: Resolver<
        Maybe<ResolversTypes['User']>,
        ParentType,
        ContextType,
        RequireFields<QueryUserArgs, 'name'>
    >;
}>;

export type UserResolvers<
    ContextType = any,
    ParentType extends ResolversParentTypes['User'] = ResolversParentTypes['User'],
> = ResolversObject<{
    email?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
    name?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
    role?: Resolver<Maybe<ResolversTypes['Role']>, ParentType, ContextType>;
    __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
}>;

export type UserWithPasswordResolvers<
    ContextType = any,
    ParentType extends
        ResolversParentTypes['UserWithPassword'] = ResolversParentTypes['UserWithPassword'],
> = ResolversObject<{
    email?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
    name?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
    password?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
    role?: Resolver<Maybe<ResolversTypes['Role']>, ParentType, ContextType>;
    __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
}>;

export type Resolvers<ContextType = any> = ResolversObject<{
    Query?: QueryResolvers<ContextType>;
    User?: UserResolvers<ContextType>;
    UserWithPassword?: UserWithPasswordResolvers<ContextType>;
}>;